move validation functions to separate files
